chore(config): consolidate environment utils
