+++ archive/services/api/bitget/ws-client.ts
@@ -13,8 +13,7 @@ import { OHLCData, Timeframe } from '../../../types/chart';
 import { OrderBookData } from '@/types/common/orderbook';
 import { ExchangeType, ProductType } from '@/types/constants/enums';
 import { logger } from '@/utils/common';
-import { IS_DEV } from '../common/environment';
-import { getApiConfig } from '../common/environment';
+import { IS_DEV, getApiConfig } from '@/config/environment';
 import { BitgetApiClientOptions } from './interfaces';
 import { formatSymbol, createWebSocket, WebSocketSendQueue } from './utils';
 import { BitgetDataTransformer } from './data-transformer';
