+++ server/bitgetWebSocketManager.ts
@@ -13,7 +13,7 @@ import WebSocket from 'ws';
 import { EventEmitter } from 'events';
 import { ExchangeType, ExchangeProductType } from '@/types/constants/enums';
 import { logger } from '@/utils/logger';
-import { getApiConfig } from '../services/api/common/environment';
+import { getApiConfig } from '@/config/environment';
 import { toBitget, fromBitget } from '../utils/timeframe';
 import { toProductType, safeExchangeType, safeProductType } from '@/utils/exchangeTypeUtils';
