+++ services/api/bitget/ws-client.new.ts
@@ -12,8 +12,7 @@ import { ProductType, BitgetCredentials } from '@/types/api';
 import { OHLCData, Timeframe } from '../../../types/chart';
 import { OrderBookData } from '../../../types/market';
 import { logger } from '@/utils/common';
-import { IS_DEV } from '../common/environment';
-import { getApiConfig } from '../common/environment';
+import { IS_DEV, getApiConfig } from '@/config/environment';
 import { BitgetApiClientOptions } from './interfaces';
 import { formatSymbol, createWebSocket } from './utils';
 import { BitgetDataTransformer } from './data-transformer';
