+++ services/errors/handler.ts
@@ -5,7 +5,7 @@
 import { toast } from 'sonner';
 import { AxiosError } from 'axios';
 import { ApiErrorHandlerOptions, WebSocketErrorHandlerOptions } from '@/types/api';
-import { IS_DEV, IS_BROWSER } from '../api/common/environment';
+import { IS_DEV, IS_BROWSER } from '@/config/environment';
 
 // エラーの種類を区別するための型
 export type ErrorSource = 'api' | 'websocket' | 'chart' | 'general';
